Threading is now included in MonteCarlo, and CylPolyRad, and BCC, as WaveMetrics has fixed teh bug in Igor in version 6.10B04
Threading can now be added to all appropriate fit functions, provided we require 6.10 (when it is a release version)

Threading issues with MonteCarlo appear to be due to NR ran() routines, which are not thread safe. Duplicate function instances (with different names) have been created. This works, but is inelegant.