Change (1): In preparation for release, updated pragma IgorVersion=6.1 in all procedures
Change (2):
As a side benefit of requiring 6.1, we can use the MultiThread keyword to thread any model function we like. The speed benefit is only noticeable on functions that require at least one integration and at least 100 points (resolution smearing is NOT threaded, too many threadSafe issues, too little benefit). I have chosen to use the MultiThread only on the XOP assignment. In the Igor code there are too many functions that are not explicitly declared threadsafe, making for a mess.