Added routines to enforce short enough names to keep the Igor-generated object names to less than 31 characters (once I add extensions). Changed the ExecuteProtocol to automatically trim text from the HFIR data file names, since they are ridiculously long.
Added routines in 1D Data loader (and XML loader) to enforce 25 characters, in the case that a long file name was generated by another program. 

If XML files are generated, "x" is tagged at the end of the suffix, as in ".ABSx".
Added menu items to each package menu for feedback, movies, and updates

Made sure that new table entry of SamplePosition did not cause compile errors with other facility templates. right now, this is NCNR-specific, but easity added to other facilities if they provide the stub. then the bits of the table could be added.