Fixed typo in the parameter description in CoreShellCylinder
W_Sigma is now appended to the table in the wrapper panel when a fit is done.
It is cleared whenever functions or data are changed.

The naming of reports is changed to help make the names more descriptive and unique.

DataSetHandling - -simple checkbox on the graph to change to a linear display
that highlights the high Q - to be sure that data oscillates around zero
after background subtraction (this doesn't have to stay)

Added a crude batch file converter to take the 5-column grasp data (read in as
if it was old-school desmeared USANS data) and write it back out in our 6-column
ASCII format. Did this for Andrea.