fix for ticket #279: calculation of the slope when desmearing.
Now the behavior is for the desmearing panel to use the slope as found when loading the .cor data. For a fresh file, it proceeds as usual. For a previously-loaded file, be sure to NOT re-load the file (<no> from the dialog) so that the slope is not recalculated, and the data is simply plotted. Then the slope information is copied along with the data to the desmearing folder.

The changed behavior is when selecting the "Extrapolate" tab. Now the extrapolation is automatically calculated from the loading parameters (not re-fit) and displayed on the graph. If the extrapolation is good, nothing needs to be done. In fact, since the extrapolation (slope) should be acceptable from the load, the mask, extrapolate, and smooth tabs can all be skipped, proceeding directly to the desmear tab.