Added new model with Yun Liu's TwoYukawa structure factor. A one-yukawa version is also added, but only in the XOP. Note that the SQ calculation is AAO and NOT threadsafe, unlike all other fitting functions. The SQ calculation has not yet been added to any form factors.
Corrected Fractal_PolyCore to return P*S rather than just P.

Only allow a 2D simulation to be saved inn 2D VAX format if it is RAW counts, not absolute rescaled data. Otherwise the DP values are mangled on conversion to VAX integers. In the future, the data may be "unscaled" if there is an outcry.