Changes to SANS reduction that apply to other Facilities:
These changes are related to the propagation of errors in 2D, on a 
per-pixel basis. These changes only affect the errors that are reported in
the QxQy ASCII file output. The 1D code is unaffected.

If these changes are not implemented, then errors of zero will be substitued as defaults
for these experimental errors.

Upon data loading, an error matrix, linear_data_error is generated and filled with 
error values appropriate for Poisson statistics (not simply sqrt(n)).

4 functions in FACILITY_DataReadWrite.ipf have been added, and they are rather
self-explanatory:
- WriteBoxCountsErrorToHeader()
- WriteTransmissionErrorToHeader()
- getBoxCountsError()
- getSampleTransError()

In FACILITY_Utils.ipf, the AttenuatorTransmission() function now returns
an additional parameter, atten_err, which is one standard deviation of the
attenuator transmission value. It returns a default error=0 (which is
correct if no attenuation is used). Facilities can fill this function in
with their own estimates for the uncertainty in the attenutator transmission.