Initial separation of facility-specific initialization from the USANS reduction procedures. This is being done to enable the KIST-USANS instrument to use the NCNR reduction scheme. KIST will duplicate the .bt5 file structure but use a different extension.
For this, files for KIST have been created for USANS-specifc, and for some SANS-specific procedures that are entangled with the XML writing and with the USANS simulator.

Corresponding changes (new U_NCNR_Utils) have been added for NCNR-specific USANS initialization.

More will need to be done to disentangle the U_CALC, but at a later date. See ticket #334.