changes to FFT routines to clean things  up for a beta release at the next startup. No functionality changes, just cleaning up the operation, interface, and menu items.
Some important changes to the polarization routines. New equations for calculating the coefficient matrix. Now appears to be correct. Also, proper proportions are used when adding multiple files together to the matrix.

Change to the Model Docs is an updated reference.

WorkFileUtils was broken when I changed to use linear_data exclusively. This has now been fixed and it operates as expected.

PackageLoader now has menu items (under Macros) for the Polarization routines. These may later be moved to the SANS menu.