changed the number of significant digits visible in the Catalog table, so that the wavelength, SDD, and sample position don't show up as odd floating point values, when they're only off by a smidgen. If they are significantly off, the real value will be seen.
Modified the file loader for Linearized fits to correctly handle files that are loaded with file names that are too long. Now they are loaded and plotted correctly, as the plot manager does. Now, when fitting, if the file name is too long (from the popup) then the user is prompted to pick the data set, since I don't know which one it is anymore, since the name was changed. Not too big of an issue. Most (All, I hope) of the file writing routines now enforce short, compliant names. this is ticket #363