Changes to the EventModeProcessing to put some of the functions on the panel as buttons, rather than on the Macros menu.
Added a menu item in the package loader to invoke the event panel. The reduction procedures must be loaded first.

Added Matt's fixes to the rescaling panel so that it will remove the "_RA" suffix when the scaling is reset to I vs q. A benefit of this is that the dependency works again (automatically) - so users could toggle between scalings and adjust coefficients if they desired.