Added the binning of QxQy data to 1D as a menu option under SANSModels->2D utilities. Mostly as a consistency check for the 2D data.
Added a package loader for the Event Mode Processing to the NCNR package loader, so that it behaves like the real spaces modeling and the polarization.

Added routines that split large event files, and then allow the accumulation of the slices from each chunk as it is processed. Eliminates the need to add a large number of split files together during the reduction.