Changes include:
updates to the simulation through SASCALC to allow easier "scripting" of SASCALC for simulation of "experiments"

updated some of the calibration values for the 10m SANS @ NGA

fixed bug in how the transmission error were being displayed in the TransFile table (bad sort). Data was not affected, only the display.