modified event loader so that it compiles correctly if the XOP is not installed
changed control polling in NSORT to explicitly look for the NSORT panel

added a testing procedure to NSORT to use only names to combine. Somewhat more "manual" in implementation than the combine panel, but very useful for combining the results from a simulation. Uses the NSORT panel in the same way ay the Combine Panel.

Changes to saving the 1D sim data to get proper header information into the file, including a proper label.

Added a macros menu for the MC Scripting beta operations. Can use an "x" to turn them off/on since they are really not for general use (yet).

Updated the simulation examples, including dry run for 2D to estimate overall run time, and 1D simulation to return a table of count rates.

Added a function stub (totally non-functional) for Write_ABSHeader_toNotebook() that in the future could be used to mine the 1D data files - to help group them, combine them, process them, and at the very least, help to identify what is in them and what was done to them.