10m SANS is now "on-center", so the tolerance for the center in ProDiv has been re-instated
All waves in the CatTable are now double precision so that the floating point details can be seen, like when the attenuator number = 2.997, rather than 3.

NCNR_utils - now rounds the attenuator number to get an integer. 2.9999 was being truncated to 2, not the "real" value of 3.

Some unfinished changes to the non-XOP 2D simulation, testing the pixel assignment.