fixed incorrect units in Beaucage model (B) in model and help file. Calculation was and still is correct.
Changed some of the FFT examples so that they are more easily used.

Added a FFT "GenericModel" that can be used with the simulator. Whatever is drawn in the matrix is used for the simulatin. There are limitations -- since FFT is used for the calculation, the q-points are interpolated and may not be sufficient for the instrumental q-range. Also, since the FFT q-range is finite, the random deviate calculation is generally unreliable. So scattering power, multiple scattering, are unreliable. Don't expect anything reasonable from the 2D Monte carlo for this reason.

The FFT panel now shows the (real) Qmin and Qmax that you'll get out.

Fixed some issues with updating the 1D graphs in the VCALC panel. Now the SlitSmeared calculations are plotting properly, although I'd like to now remove the T/B detectors since there's not enough data there.