Correcting the search in the patch panel for certain cases where multiple run numbers could be returned if they matched the year.
Adding routines and preferences to be able to add together raw data files with different attenuation. Not sure why this was not possible in the past. There must be a good reason for this. This is toggled on/off with a SANS preference checkbox. Default is OFF

Adding to Transmission "guessing" to show buttons to bump the guess  to one more or one fewer character, rather than forcing an exit if the match is not correct.