Updating the patch panel for the VSANS variables, a first pass at including the essential values, and a section that has per-detector values.
Still needs to be fully stress tested to be sure that it does all of read/write operations correctly.

adjusted the "last file loaded" global variable that is displayed on the data display to only update when a file is loaded for display, not when any file is loaded for the catalog, or for the patch panel.