Added procedures to increase the number of elements in a reduction protocol, to allow the trimming points to be added directly to the protocol.
The protocol can also now be written directly to the reduction block of the data file. This can serve as a record of what was used, and also for multiple reduction.

Although the protocol is not a NICE -generated block in the HDF file, there does not appear to be any issues with writing out this new wave to the data set, as long as I handle the error if I try to read it (since i't not there to begin with).

Protocol panel is now aware of this change, and the multiple reduce panel has been updated and added to the Inlcudes list.