more fixes to DIV file generation using the high-resolution detector
added extra header lines to the the QxQy ASCII output format. there are now 30 header lines, and 7 or so are unused to allow for future additions.

the default mask for the high res detector has been enlarged to account for the large beamstop that is used during collection of the sensitivity file.