changed the file extension on VSANS QxQyASCII output to have the panel tag in th e middle of the name to keep the .DAT extension visible for sasView
Changed the behavior of the VSANS averaging protocol. When "saving file" is selceted, choices are either to concatenate or individual. concatenate has NOT been changed. Individual save will now save each individual detector panel data in as
eparate file with the detectr panel string tagged on the end. This replaces the
tedious procedure of multiple mask files an multiple reduction passes.

The save as ITX function is still available through a menu option but is current
ly not an option for a protocol.