added comments in NCNR_Utils identifying the proper instrument prefix values for each instrument
Saving individual data sets from VSANS will now properly ignore the back detector (if preference set) and will handle all of the current binning types.

In the preferences, the bin width (as a multiple of pixels) is now active. This setting is a multiple of the natural deltaQ of one tube width.