fixed potential bug in mask panel for Igor 9 (beta) (ShowTools/A)
changed the behavior of the ShowAvgMask to whow the effect of both the "regular"
mask and the annular/sector mask so that the actual range that contributes
is shown, not just part.

changed the color scale to be per-carriage rather than per-panel. Visually
looks a lot better now - where the LR panels are closed, it looke like
a single detector now. Not sure why I didn't do this before.