changes to polarization routines: 1-possible fix of math error in polarization matrix. still in progress 2-updating calcualtions in VAX routines to not changes in calculation 3-updating calculations in Nexus-SANS so that they work properly with the Nexus structure 4-updating calculations in Nexus-VSANS to mirror SANS calculations
in SANS and VSANS with Nexus files, the handling of folders with uu,du, etc. and the location
of the polarization matrix waves is different than VAX calculations (due to
different handling of the Nexus files)

Temporary fix for an issue where the VAX-> Nexus conversion kept the VAX time
format which was parsed incorrectly when expecting a Nexus file.
If mis-parsed, then it is assumed it is trying to read VAX format.
All appears to be correct now.