more changes to the polarization routines: -SANS Nexus should be complete at this point - VSANS still needs work - changes that altered non-polarization ipf files have been reversed so that they do not impact the regular reduction process.
all appears to work with SANS NExus/polarization through some basic testing
of example data sets. More testing is needed and finalization of the
question of the polarization matrix coefficients. Currently the
revised coefficients are in place based on the corrections to the
original reference. (the changes seem to have little effect).