Updating BT5 Loader
Updating the BT% loader to be more generic, testing each of the 5 (or 6) columns to see which one is the count time. Before, I was assumning that the count time was either column 2 or col 3. To be sure that nothing will be broken when/if NICE changes, explicitly test each time