update to zero point table selection
Updated to progammatically determine if the zero point tables are needed by finding the average of the first row of teh calibration wave. if not the "perfect" value, then the zero point waves are not needed, otherwise use the tables.

not a perfect method, but should eliminate mistakes of not using the tables. only a significant issue for beam centroids with the graphite monochromator