Updated QxQy display of NXcanSAS 2D data
Updated the display of NXcanSAS 2D data to properly display in QxQy space, not what was essentially an even pixel grid with Q-labels.

Data is now interpolated to the overall QxQy range. Each panel is interpolated and added to the overall image separately. There can be some distortions, but I have tried to minimize these

Also added a menu item to save WORK folders as NXcanSAS 2D, without the need to set up a dummy protocol.

Changed the file filter on the NXcanSAS 2D plot panel to only display NXcanSAS data files, not any of the QxQy ASCII files (or anything else that might be present, only the NXcanSAS Files)