Added Beam Center Delta for Narrow Slit Data
Added the newly determined values for MLx and FLx delta values for Narrow Slit collimation, which are different than delta values for pinhole collimation.

Updated the logic to add a switch based on the collimation (number of guides field) to use the NS values only if NARROW_SLITS is returned.

All other cases use the pinhole delta values.